pyproject.toml
# pyproject.toml
# /var/www/drecall/pyproject.toml

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drecall"
version = "0.1.0"
description = "Tech-agnostic cognitive RAG system for document and chat history"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DRecall Contributors"}
]

# All dependencies are required, but the architecture is tech-agnostic via interfaces
# You can swap implementations by changing config.yaml (wrapper pattern)
dependencies = [
    # CLI & Core
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",

    # Normalization (swappable via INormalizer interface)
    "markitdown>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",

    # AI/ML (swappable via IClassifier, IVectorizer interfaces)
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",

    # Storage (swappable via IStorage interface)
    "sqlite-vec>=0.1.0",
]

[project.optional-dependencies]
# Development dependencies only
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
drecall = "drecall.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "drecall*"]

[tool.setuptools.package-data]
drecall = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
